Switch to list concatenation for variables specification.
Add the list of requested variables to the `args` list by concating.
Previously I was manually creating the space seperated list (i.e string)
when a list of variables was passed to the girdded inerpolator function.
This space seperated string was not correctly parsed by the
`mpas_tool.logging.check_call` function resulting in the variables not
being inerpolated.